Test suite for unit and integration tests
