+++ bors.toml
@@ -1,4 +1,4 @@
 status = [
     "Test",
 ]
-delte_merged_branches = true
+delete_merged_branches = true
