+++ src/vga_buffer.rs
@@ -85,10 +85,10 @@ impl Writer {
                 if self.column_position >= BUFFER_WIDTH {
                     self.new_line();
                 }
-                
+
                 let row = BUFFER_HEIGHT - 1;
                 let col = self.column_position;
-                
+
                 let color_code = self.color_code;
                 self.buffer.chars[row][col].write(ScreenChar {
                     ascii_character: byte,
@@ -98,7 +98,7 @@ impl Writer {
             }
         }
     }
-    
+
     // Writes the given ASCII string to the buffer
     // Wraps lines at `BUFFER_WIDTH`. Supports the `\n` newline character. Doesn't support
     // strings with non-ASCII characters, since they can't be printed in VGA text mode
@@ -110,7 +110,7 @@ impl Writer {
             }
         }
     }
-    
+
     // Shifts all lines on line up and clears last row
     fn new_line(&mut self) {
         for row in 1..BUFFER_HEIGHT {
@@ -122,7 +122,7 @@ impl Writer {
         self.clear_row(BUFFER_HEIGHT - 1);
         self.column_position = 0;
     }
-    
+
     // Clears a row by overwriting it with blank characters
     fn clear_row(&mut self, row: usize) {
         let blank = ScreenChar {
